PN-12663: refactored useB2B annotation
