+++ pom.xml
@@ -1689,14 +1689,14 @@
                             <typeMappings>
                                 <typeMapping>OffsetDateTime=java.lang.String</typeMapping>
                             </typeMappings>
-                            <inputSpec>https://raw.githubusercontent.com/pagopa/pn-delivery-push/48ccb5e03e404b1895bbbdb16e83890de8b74dee/docs/openapi/api-external-pn-delivery-push-b2b-pg.yaml</inputSpec>
+                            <inputSpec>https://raw.githubusercontent.com/pagopa/pn-delivery-push/refs/heads/develop/docs/openapi/api-external-pn-delivery-push-b2b-pg.yaml</inputSpec>
                             <generatorName>java</generatorName>
                             <library>resttemplate</library>
                             <generateApiTests>false</generateApiTests>
                             <generateModelTests>false</generateModelTests>
                             <configOptions>
-                                <apiPackage>it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypush2b.api</apiPackage>
-                                <modelPackage>it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypush2b.model</modelPackage>
+                                <apiPackage>it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypushb2b.api</apiPackage>
+                                <modelPackage>it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypushb2b.model</modelPackage>
                                 <dateLibrary>java8</dateLibrary>
                                 <annotationLibrary>none</annotationLibrary>
                                 <documentationProvider>source</documentationProvider>
@@ -1707,8 +1707,8 @@
                                     @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)
                                 </additionalModelTypeAnnotations>
                             </configOptions>
-                            <apiPackage>it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypush2b.api</apiPackage>
-                            <modelPackage>it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypush2b.model</modelPackage>
+                            <apiPackage>it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypushb2b.api</apiPackage>
+                            <modelPackage>it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypushb2b.model</modelPackage>
                         </configuration>
                     </execution>
