+++ src/main/java/it/pagopa/pn/client/b2b/pa/service/IB2BDeliveryPushServiceClient.java
@@ -1,6 +1,6 @@
 package it.pagopa.pn.client.b2b.pa.service;
 
-import it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypush2b.model.LegalFactDownloadMetadataResponse;
+import it.pagopa.pn.client.b2b.generated.openapi.clients.deliverypushb2b.model.LegalFactDownloadMetadataResponse;
 import it.pagopa.pn.client.b2b.pa.service.utils.SettableBearerToken;
 import org.springframework.web.client.RestClientException;
